payment_app/fb_functions/requirements.txt
pycryptodome==3.9.7
phpserialize==1.3
google-cloud-firestore==1.7.0
